feat: add index manager
