+++ eincraft/__init__.py
@@ -1,3 +1,5 @@
 from .core import EinTen
+from .tensor import EinTenBaseTensor
+from .contraction import EinTenContraction, IndexMap
 
-__all__ = ["EinTen"]
+__all__ = ["EinTen", "EinTenBaseTensor", "EinTenContraction", "IndexMap"]
